requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.7
SQLModel==0.0.22
psycopg[binary]==3.2.1
python-dotenv==1.0.1
yt-dlp==2024.10.22
pydantic-settings==2.6.1
httpx==0.27.2
openai==1.51.0
google-search-results==2.4.2
